Add function to choose based on benchmark data
I've thought long and hard about this one. This is a function that is supposed to choose an algorithm based on parameters and benchmarks of performance of each algorithm. The benchmark data isn't there yet. There will be a CMake target to create them for the current machine. Let's first implement this choose function though. Quite complex, all of this.