Make more passes if local memory is a bottleneck
This will most likely be the case when working on a CPU, which has huge global memory but small L1 cache, and comparatively few cores. We need to make more passes then because the local memory can't hold all the intermediary values all at once. This local memory needs to hold the area contributions of each vertex (which is coord_t per vertex and each vertex is coord_t*2, so half of the memory required for the vertices).