Add benchmarker application
This won't be working properly yet, since the benchmarker must provide data about the current CPU and GPU set-up as well. It must also output the data in such a way that it disambiguates when ran on different set-ups. But first we'll need to solve a bug that causes the library to crash with a ParallelogramException upon encountering polygons with 100 or more vertices.