Don't forget the last iteration of aggregate sum
When the size ends up at 2, the offset is still 1 and the sum for-loop will stop. We need to do one last iteration to sum the last two numbers together. This only gets executed by worker 0, so we can do it after the barrier and save one extra barrier that way as well.