Add BenchmarkData header
The purpose of this header is to group all benchmark data together. In the future we will probably want CMake to add stuff to this file so that someone who compiles Parallelogram can include benchmark data for his own computer easily. For now he must hardcode his includes in here.