Round global memory buffer size up to size per work group
The result may not be added up completely, but each of these workers is going to request memory from that buffer. And if the buffer size is not a multiple of the work group size that means that the worker is accessing memory outside of the buffer. On a GPU that doesn't matter but on the CPU that causes a segmentation fault.