Add script to find Eigen and its version
It's rather convoluted right now. I plan to include the GetEigenVersion inside the script, just inline it for making it easier to copy-pasta. There's also some stuff we could do better with ExternalProject_Add instead of the plain download we're doing now, but we need to do the plain download anyway in order to get the version number...