Download via ExternalProject_Add
This is how CMake is supposed to work. It makes the configure step much faster, and allows building the main library completely before Eigen is even touched since Eigen is only necessary for the benchmarking application. Much more lazy. Such fast. Only annoying thing is that you'd normally want to disable all the stuff that Eigen does in its CMake script because you only want the headers. That configure step is HEAVY and doesn't cache well, so it needs to execute stuff every time you build the benchmarker. And that increases the build time of that little application from 5 seconds to a minute or so.