Compute interpolation polynomial
This polynomial is intended to be used when a device is used that we don't have benchmarks for, in order to get a prediction of how long the algorithm would run. This then allows us to select the best device to run the algorithm on.
Right now this is a bit messy. It is not yet written to extend to other tests than the one area test we have now and the interpolation is strictly linear. But that's what we will improve next.