Only depend on GoogleTest when building it from source
This is to make sure that GoogleTest is already built before the tests are built.
If GoogleTest is installed on the system, then this is unnecessary and outputs dev warnings in the console log of the user trying to build these tests that the target GoogleTest doesn't exist. We should find some solution for that in the Find script I guess, but I wouldn't know how.