Query host device information if host is not OpenCL
On some computers, the host CPU doesn't support OpenCL (computers like my laptop). So for the host, we may not use OpenCL to query device information. We must query it manually.
This is an implementation for Linux. On other systems, it will say that the CPU identity is unknown and will not output any device information. I'll create a Windows implementation shortly. I don't have a way to test anything on MacOS though.