Add class to factor out benchmarking code
I'd hate to write that whole benchmarking code out for every method we're going to be testing. So let's make a factory class that has most of that code, as far as we can factor it out. It's going to need quite a lot of input though, but in the long run it'll be easier.