Add function to run one test as parameter
This is sort of cheating, but this function can be a simple lambda; very short. All it needs to do is call the function that we need to benchmark with the appropriate parameters. Since the methods we're benching are so varied in the input parameters that they take, we can't make a one-size-fits-all type and hope that we can generate the correct input parameters for the function. So instead we'll accept a lambda that calls the benchmarked function with correct parameters. So then if you wish to test, for instance, polygon offsets, then you need to provide a lambda that accepts one polygon, and calls the offset method on it with a certain constant offset distance. It won't be all-powerful, but it will suffice for a while.