Statically define all benchmarks
I think eventually we'll need to move this to a function if we are ever going to use the factory pattern of SimplePolygonBenchmark to have a non-default value for any of those fields. We'll see about that when we get there.
This also requires adding the device to the parameters of the benchmarks, so that we don't need a copy for each of the devices. We don't know what all of the devices are, at this point.