Dynamically store benchmark data under test name
Instead of creating a new unordered_map for each test, we now store it all in the same dictionary so that we can dynamically add more tests without having to define a new unordered_map in BenchmarkData.
Currently this doesn't compile due to double definitions in BenchmarkData.h. We'll address that shortly.