Don't return anything from compute_interpolation
Semantically it would make sense to return the interpolation results from this function, but since the caller would then have to interpret this arbitrary vector of doubles, I'd rather not do it that way and just keep the meaning of the doubles in this vector contained in the compute_interpolation function.

Eventually we'll have to change this (back). When choosing a strategy and the current device has no measurement, we'll have to fill in the actual system information in order to get the time estimates and make a choice of strategy. We'll see what to do about this then.