Load benchmark data via singleton constructor
This way the benchmark data needs to be loaded only once, not for every single function call.
This also guarantees that the benchmark data is loaded. Previously you could forget the load_benchmarks() call before using the data. Now the data is only available if you have the singleton instance of BenchmarkData, and once you make an instance the benchmark data will load. Safer