Fix edge cases and handle horizontal edges
The previous edge case handling was too naive in the case where multiple edges overlap and then hit the requested point.
Also implemented handling of horizontal edges. This is all quite arcane and I'm not 100% sure that I understand it completely either. We must rely on the tests to make sure that all edge cases are hit. If we miss some edge cases then let's hope that they are not common enough in real life scenario's.