Remove separate edge counting again
It didn't work in all cases. Turns out that some cases that this was trying to solve are just ambiguous.
The case that this was intended to solve concerned a line tracing back over itself. This is ambiguous since the direction of the polygon is neither clockwise nor counterclockwise or both, depending on how you see it. So our convention of counting edge cases inclusive only for clockwise winding breaks down there.