Limit vertices per work group by memory
It's probably not a problem for this particular algorithm, since each work item only uses 8 bytes. Since the minimum local memory size is 32kB, you'd need to have 4000 processors running on that minimum amount of memory. No GPU exists that is so out of balance.
However, this also serves as the basis for more complex algorithms. I'll probably be factoring this code out to a more central place so that it can be reused for other algorithms too.