Fix type of boolean include_edge
There is no boolean type in OpenCL C, so we must use an int there. It was passing a boolean as argument though, hoping it would get cast to cl_int in the process. It doesn't though... This causes a misalignment of all parameters and eventually causes the output to be undefined memory. It's as if the kernel does nothing at all and just leaves the input as it was: arbitrary memory.