Fix clamping algorithm
This was intended to perform better than a real clamp since we just need the sign. However it's actually wrong since modulo 1 on an integer will always produce 0. We'll have to do a clamp for now until we can come up with something slightly better (but that would be micro-optimisation).