Simplify hit check and correct edge cases
The hit check in here no longer uses the trick with underflowing unsigned integers in order to prevent a branch. It was not, in fact, preventing a branch... It was preventing another boolean check in the same if-statement which is unnecessary since it's included in the comparison operator if we were to use that unsigned integer trick. However this was also preventing us from getting the correct result with the edge case of a point being at the height of a vertex in the coordinate, where we must include the point in one of the edges but not the other. So we're taking the performance hit now. The performance impact should be very slim since this essentially just adds one integer comparison and one boolean 'and'. On a GPU core with full threading this should be no problem at all.