Remove branch for rising vs. falling edges
This is now done in one branch by multiplying with -1 or xoring if it's the falling edge instead of the rising one. This should save some of the code points, and eventually save a good bit of calculation then that is ignored due to being on an inactive branch. GPUs be weird like that, yo.