Add batch processing class for SimplePolygon
It's empty right now. No functions to speak of.
This class is supposed to prevent code duplication between SimplePolygon and Polygon. Hopefully, many methods of SimplePolygon can be reduced to calls to a batch processor when supplying an iterator over a single element: itself. And then the Polygon class can call the same function to get an answer for multiple simple polygons.
And this also helps towards the goal of increasing performance with a lot of data for the user, since the user can batch up some of its processing.