Store total size of all polygons in batch
We need this to use our learner-trained strategy chooser to select the most optimal algorithm to perform operations with.
Word is still out on whether the iterators should be const. We might not be able to perform modifying operations on the polygons if they are. But if they are not, we might not be able to use this class on const polygons.