Add function to divide edges over work groups and passes
This is the complex logic that was the heart of the optimisation that the previous implementations were doing. Hopefully this version be re-used in many of them, perhaps adjusted for algorithms that work on loose vertices or different combinations of data in the polygon. And maybe it can be tested more easily as well, even though it is a private function and requires some mocking of sorts to get the proper device statistics in.