Rename Parallelogram to Apex
A much cooler name, really. I already know a logo and everything. This name is taken by a semi-well-known project by Apache, but libApex seems to not be a household name so I think we're fine. The name 'Apex' has a meaning in geometry as the summit of a cone or pyramid, or the tipping point of a curve.