Remove divide_edges and replace with plan of subbatches
If we're going to implement an algorithm that batches can stay in device memory for longer, then we'll need a different approach. Using this approach, at least one subbatch will be able to stay in memory, saving some time later on. Also, if we are able to combine multiple operations on a batch we might be able to do that on a per-subbatch basis simply by passing the call on to each of the subbatches. This mostly just documents the idea so that it's registered while I'm implementing it; so I don't forget the idea.