Add dummy target to build if gtest is found
We need to ensure that GoogleTest is built before building any tests, if it needs to be built. Whether or not it needs to be built is completely transparent inside FindGoogleTest so there is no way to know from the outside. However we still need to make the test targets depend on the GoogleTest build target if it needs to be built, so if it doesn't need to be built, we now create a dummy target that you can 'build' always even if GoogleTest is just found on your system.