Clear subbatches if failed to re-batch
We don't want to leave it in a state where the subbatches don't add up to all of the polygons. At least leave it in a state where there are no subbatches then, so that we can properly detect that it has not been prepared to load into global memory yet.