Add field to remember if this batch is loaded on a device's memory
This will give us trouble later on, I'm betting, if the user is working with multiple batches and some of them are moved and some of them are copied and the data is modified in some of them but not in others. That could get real complex especially if we can't detect that the data has been modified externally, such as in a copy of a batch. Maybe we'll need to reset is_loaded when copying. We'll cross that bridge when we get there I guess.