Move template Device class to only constructor
The entire class doesn't need to be templated. Just the constructor.
I don't know why that bit in OpenCLDevices.cpp was now necessary. It wouldn't accept the constructor with parameters, nor without parameters. But it needs to be a pointer to cl::Device* anyway so let's keep it at that.