Add test for loading an empty batch in memory
The creation of this test required all of the mocks that were committed right before this commit. Quite an exercise.
The test also revealed another mistake that I made. I can't verify the actual contents of the buffer, because that buffer was a local variable in the load() function. It actually gets destroyed at the end of the function, undoing all of the loading and making the entire function moot (and preventing me from testing it). We're going to need to store this buffer somewhere alongside the loaded_in_memory thingamajig.