Actually resize vector instead of just reserving memory
Reserving memory gives a guarantee that we don't get segfaults from reading or writing there, but it's still nondeterministic what data we read from it. Now we can be certain that it'll be filled with 0's. This has no relation with how it works in OpenCL (where it's nondeterministic) but it's good to have deterministic tests.