Add test for caching batches on device
When we don't need to reload it because it's already on the device.
There is a serious issue now with the memory usage when multiple batches are on the device. We're unable to detect this. Maybe we need to create a wrapper class for cl::Device that tracks all of the batches that are on it...